feat(axum): separate portfolio route implementation from openapi spec
